requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
pytest==7.2.0
requests==2.28.1
